+++ themes/ideaspace-360-photo-tour/config.php
@@ -4,8 +4,8 @@ return [
 
   '#theme-name' => 'IdeaSpace 360 Photo Tour',
   '#theme-key' => 'ideaspace-360-photo-tour',
-  '#theme-version' => '1.1',
-  '#ideaspace-version' => '>=1.0.2',
+  '#theme-version' => '1.2',
+  '#ideaspace-version' => '>=1.1.0',
   '#theme-description' => 'Create and publish a 360 photo sphere tour with info hotspots. Attach text annotations to hotspots.',
   '#theme-author-name' => 'IdeaSpaceVR',
   '#theme-author-email' => 'info@ideaspacevr.org',