+++ themes/ideaspace-3d-model/config.php
@@ -4,8 +4,8 @@ return [
 
   '#theme-name' => 'IdeaSpace 3D Model',
   '#theme-key' => 'ideaspace-3d-model',
-  '#theme-version' => '1.0',
-  '#ideaspace-version' => '>=1.0.2',
+  '#theme-version' => '1.1',
+  '#ideaspace-version' => '>=1.1.0',
   '#theme-description' => '3D model viewer with info hotspots. Attach text annotations to hotspots.',
   '#theme-author-name' => 'IdeaSpaceVR',
   '#theme-author-email' => 'info@ideaspacevr.org',