+++ themes/vr-view-360-image/config.php
@@ -4,8 +4,8 @@ return [
 
   '#theme-name' => 'VR View 360 Image with Hotspot Navigation',
   '#theme-key' => 'vr-view-360-image-hotspot',
-  '#theme-version' => '1.0',
-  '#ideaspace-version' => '>=1.0.1',
+  '#theme-version' => '1.1',
+  '#ideaspace-version' => '>=1.1.0',
   '#theme-description' => 'Photo sphere viewer with navigation hotspots. Based on the "VR View for the Web" by Google.',
   '#theme-author-name' => 'IdeaSpaceVR',
   '#theme-author-email' => 'info@ideaspacevr.org',